requirements.txt
Django==3.1.3
requests
pyjwt
django-cors-headers
djangorestframework
mysqlclient
oauthlib
requests-oauthlib
gunicorn
money
babel
django-multiselectfield
pandas
django-extensions